# pwnedapi/static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PwnedAPI",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://api.pwnedhub.com",
      "variables": {}
    }
  ],
  "paths": {
...
    "/notes": {
      "get": {
        "tags": [
          "Misc"
        ],
        "summary": "Note Read",
        "operationId": "NoteRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Misc"
        ],
        "summary": "Note Update",
        "operationId": "NoteUpdate",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteUpdateRequest"
              },
              "example": {
                "content": "##### Welcome to PwnedHub 2.0!\n\nA collaborative space to conduct hosted security assessments.\n\n**Find flaws.**\n\n* This is your notes space. Keep your personal notes here.\n* Leverage popular security testing tools right from your browser in the tools space.\n\n**Collaborate.**\n\n* Privately collborate with coworkers in the messaging space.\n* Join public rooms in the messaging space to share information and socialize.\n\n**On the Move**\n\n* PwnedHub 2.0 is built with mobility in mind. No need for a separate app!\n\nHappy hunting!\n\n\\- The PwnedHub Team\n\n"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/tools": {
      "get": {
        "tags": [
          "Misc"
        ],
        "summary": "Tools Read",
        "operationId": "ToolsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "Tool Create",
        "operationId": "ToolCreate",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolCreateRequest"
              },
              "example": {
                "name": "ID",
                "path": "id",
                "description": "Displays the user ID of the current shell user."
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/messages": {
      "get": {
        "tags": [
          "Misc"
        ],
        "summary": "Messages Read",
        "operationId": "MessagesRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "Message Create",
        "operationId": "MessageCreate",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreateRequest"
              },
              "example": {
                "comment": "Hello everyone!",
                "room_id": 1
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NoteUpdateRequest": {
        "title": "NoteUpdateRequest",
        "required": [
          "content"
        ],
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          }
        },
        "example": {
          "content": "##### Welcome to PwnedHub 2.0!\n\nA collaborative space to conduct hosted security assessments.\n\n**Find flaws.**\n\n* This is your notes space. Keep your personal notes here.\n* Leverage popular security testing tools right from your browser in the tools space.\n\n**Collaborate.**\n\n* Privately collborate with coworkers in the messaging space.\n* Join public rooms in the messaging space to share information and socialize.\n\n**On the Move**\n\n* PwnedHub 2.0 is built with mobility in mind. No need for a separate app!\n\nHappy hunting!\n\n\\- The PwnedHub Team\n\n"
        }
      },
...
      "ToolCreateRequest": {
        "title": "ToolCreateRequest",
        "required": [
          "name",
          "path",
          "description"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "example": {
          "name": "ID",
          "path": "id",
          "description": "Displays the user ID of the current shell user."
        }
      },
...
      "MessageCreateRequest": {
        "title": "MessageCreateRequest",
        "required": [
          "comment",
          "room_id"
        ],
        "type": "object",
        "properties": {
          "comment": {
            "type": "string"
          },
          "room_id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "comment": "Hello everyone!",
          "room_id": 1
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}